+++ src/activities/Elsa.Activities.Http/Elsa.Activities.Http.csproj
@@ -27,7 +27,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="LanguageExt.Core" Version="3.3.29" />
         <PackageReference Include="Microsoft.AspNetCore.DataProtection" Version="3.0.0" />
         <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
         <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
