--- src/core/Elsa.Abstractions/Services/IScopedWorkflowInvoker.cs
@@ -1,7 +1,0 @@
-namespace Elsa.Services
-{
-    internal interface IScopedWorkflowInvoker : IWorkflowInvoker
-    {
-        
-    }
-}
