+++ src/core/Elsa.Core/Metadata/ActivityDescriptor.cs
@@ -14,11 +14,11 @@ namespace Elsa.Metadata
 
         public string Type { get; set; }
         public string DisplayName { get; set; }
-        public string? Description { get; set; }
-        public string? RuntimeDescription { get; set; }
+        public string Description { get; set; }
+        public string RuntimeDescription { get; set; }
         public string Category { get; set; }
-        public string? Icon { get; set; }
-        public object? Outcomes { get; set; }
+        public string Icon { get; set; }
+        public object Outcomes { get; set; }
         public ActivityPropertyDescriptor[] Properties { get; set; }
     }
 }
