+++ src/core/Elsa.Core/Metadata/ActivityPropertyDescriptor.cs
@@ -2,7 +2,7 @@ namespace Elsa.Metadata
 {
     public class ActivityPropertyDescriptor
     {
-        public ActivityPropertyDescriptor(string name, string type, string label, string? hint = null, object? options = null)
+        public ActivityPropertyDescriptor(string name, string type, string label, string hint = null, object options = null)
         {
             Name = name;
             Type = type;
@@ -14,7 +14,7 @@ namespace Elsa.Metadata
         public string Name { get; }
         public string Type { get; }
         public string Label { get; }
-        public string? Hint { get; }
-        public object? Options { get; }
+        public string Hint { get; }
+        public object Options { get; }
     }
 }
