+++ src/core/Elsa/ElsaServiceCollectionExtensions.cs
@@ -3,9 +3,7 @@ using Elsa;
 using Elsa.Activities.ControlFlow.Extensions;
 using Elsa.Activities.UserTask.Extensions;
 using Elsa.Activities.Workflows.Extensions;
-using Elsa.Scripting.JavaScript;
 using Elsa.Scripting.JavaScript.Extensions;
-using Elsa.Scripting.Liquid;
 using Elsa.Scripting.Liquid.Extensions;
 
 // ReSharper disable once CheckNamespace
