+++ src/scripting/Elsa.Scripting.JavaScript/Services/JavaScriptExpressionEvaluator.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
